Working on checkers for jumping
