+++ Assets/Core/Scripts/Checkers.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 26b65d5f1a972434eabd544a7dc3022b
+guid: aa711327dfe0cdd429f65ca588658fe5
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
