+++ Assets/Core/Scripts/Checkers/IsObstacleAboveChecker.cs
@@ -1,0 +1,21 @@
+using UnityEngine;
+
+public class IsObstacleAboveChecker : MonoBehaviour, IInitializable<PositionSummer>
+{
+    private PositionSummer _positionSummer;
+
+    public void Initialize(PositionSummer positionSummer)
+    {
+        _positionSummer = positionSummer;
+    }
+    
+    public void IsObstaclesAbove()
+    {
+        // return Physics.CheckSphere
+        // (
+        //     _positionSummer.SumPosition(transform.position, _positionOfCheck),
+        //     _radiusOfCheck,
+        //     _toStandLayer
+        // );
+    }
+}
