+++ Assets/Core/Scripts/Checkers/IsObstacleAboveChecker.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 72b96d299831b67439c2cc7afd34c847
