+++ Assets/Core/Scripts/Checkers/PositionSummer.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+public class PositionSummer : MonoBehaviour
+{
+    public Vector3 SumPosition(Vector3 center, Vector3 checkingPosition)
+	{
+		return new Vector3
+		(
+			transform.localPosition.x + checkingPosition.x,
+			transform.localPosition.y + checkingPosition.y,
+			transform.localPosition.z + checkingPosition.z
+		);
+	}
+}
