+++ Assets/Core/Scripts/Checkers/PositionSummer.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: de6bf7e90391211409f12bfd849c6562
